requirements.txt
altair==5.1.1
annotated-types==0.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
exceptiongroup==1.1.3
fake-useragent==1.2.1
filelock==3.12.4
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.17.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
googlesearch-python==1.2.3
greenlet==3.0.1
gspread==5.11.2
h11==0.14.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
murmurhash==1.0.10
mysql-connector==2.2.9
numpy==1.26.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pandas==2.1.1
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.9
protobuf==4.24.3
py3-validate-email==1.0.5.post1
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
selenium==4.12.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
streamlit==1.27.0
tenacity==8.2.3
thinc==8.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==2.0.4
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.2
websockets==11.0.3
Werkzeug==3.0.1
wsproto==1.2.0
zcrmsdk==3.1.0
zipp==3.17.0
